+++ negmas/utilities.py
@@ -65,6 +65,7 @@ from negmas.outcomes import (
 if TYPE_CHECKING:
     from negmas.outcomes import OutcomeRange, Outcome
 
+
 __all__ = [
     "UtilityDistribution",
     "UtilityValue",
@@ -85,8 +86,11 @@ __all__ = [
     "normalize",
     "JavaUtilityFunction",
     "RandomUtilityFunction",
+    "INVALID_UTILITY",
 ]
 
+# constants
+INVALID_UTILITY = -1000.0
 
 # Helper Types just used to make type hinting more readable
 OutcomeUtilityMapping = Union[
@@ -133,7 +137,7 @@ class UtilityFunction(ABC, NamedObject):
         self,
         name: Optional[str] = None,
         ami: AgentMechanismInterface = None,
-        reserved_value: Optional[UtilityValue] = 0.0,
+        reserved_value: Optional[UtilityValue] = INVALID_UTILITY,
     ) -> None:
         super().__init__(name=name)
         self.reserved_value = reserved_value
@@ -1169,7 +1173,7 @@ class ExpDiscountedUFun(UtilityFunction):
         beta: Optional[float] = None,
         factor: Union[str, Callable[["AgentMechanismInterface"], float]] = "step",
         name=None,
-        reserved_value: Optional[UtilityValue] = 0.0,
+        reserved_value: Optional[UtilityValue] = INVALID_UTILITY,
         dynamic_reservation=True,
     ):
         super().__init__(name=name, reserved_value=reserved_value, ami=ami)
@@ -1239,7 +1243,7 @@ class LinDiscountedUFun(UtilityFunction):
         factor: Union[str, Callable[["AgentMechanismInterface"], float]] = "step",
         power: float = 1.0,
         name=None,
-        reserved_value: Optional[UtilityValue] = 0.0,
+        reserved_value: Optional[UtilityValue] = INVALID_UTILITY,
         dynamic_reservation=True,
     ):
         super().__init__(name=name, reserved_value=reserved_value, ami=ami)
@@ -1297,7 +1301,7 @@ class ConstUFun(UtilityFunction):
         self,
         value: float,
         name=None,
-        reserved_value: Optional[float] = 0.0,
+        reserved_value: Optional[float] = INVALID_UTILITY,
         ami: AgentMechanismInterface = None,
     ):
         super().__init__(name=name, reserved_value=reserved_value, ami=ami)
@@ -1443,7 +1447,7 @@ class LinearUtilityAggregationFunction(UtilityFunction):
         ],
         weights: Optional[Union[Mapping[Any, float], Sequence[float]]] = None,
         name: Optional[str] = None,
-        reserved_value: Optional[UtilityValue] = 0.0,
+        reserved_value: Optional[UtilityValue] = INVALID_UTILITY,
         ami: AgentMechanismInterface = None,
     ) -> None:
         super().__init__(name=name, reserved_value=reserved_value, ami=ami)
@@ -1649,7 +1653,7 @@ class MappingUtilityFunction(UtilityFunction):
         mapping: OutcomeUtilityMapping,
         default=None,
         name: str = None,
-        reserved_value: Optional[UtilityValue] = 0.0,
+        reserved_value: Optional[UtilityValue] = INVALID_UTILITY,
         ami: AgentMechanismInterface = None,
     ) -> None:
         super().__init__(name=name, reserved_value=reserved_value, ami=ami)
@@ -1789,7 +1793,7 @@ class NonLinearUtilityAggregationFunction(UtilityFunction):
         issue_utilities: MutableMapping[Any, GenericMapping],
         f: Callable[[Dict[Any, UtilityValue]], UtilityValue],
         name: Optional[str] = None,
-        reserved_value: Optional[UtilityValue] = 0.0,
+        reserved_value: Optional[UtilityValue] = INVALID_UTILITY,
         ami: AgentMechanismInterface = None,
     ) -> None:
         super().__init__(name=name, reserved_value=reserved_value, ami=ami)
@@ -1966,7 +1970,7 @@ class HyperRectangleUtilityFunction(UtilityFunction):
         ignore_issues_not_in_input=False,
         ignore_failing_range_utilities=False,
         name: Optional[str] = None,
-        reserved_value: Optional[UtilityValue] = 0.0,
+        reserved_value: Optional[UtilityValue] = INVALID_UTILITY,
         ami: AgentMechanismInterface = None,
     ) -> None:
         super().__init__(name=name, reserved_value=reserved_value, ami=ami)
@@ -2034,7 +2038,7 @@ class NonlinearHyperRectangleUtilityFunction(UtilityFunction):
         mappings: OutcomeUtilityMappings,
         f: Callable[[List[UtilityValue]], UtilityValue],
         name: Optional[str] = None,
-        reserved_value: Optional[UtilityValue] = 0.0,
+        reserved_value: Optional[UtilityValue] = INVALID_UTILITY,
         ami: AgentMechanismInterface = None,
     ) -> None:
         super().__init__(name=name, reserved_value=reserved_value, ami=ami)
@@ -2072,7 +2076,7 @@ class ComplexWeightedUtilityFunction(UtilityFunction):
         ufuns: Iterable[UtilityFunction],
         weights: Optional[Iterable[float]] = None,
         name=None,
-        reserved_value: Optional[UtilityValue] = 0.0,
+        reserved_value: Optional[UtilityValue] = INVALID_UTILITY,
         ami: AgentMechanismInterface = None,
     ):
         super().__init__(name=name, reserved_value=reserved_value, ami=ami)
@@ -2132,7 +2136,7 @@ class ComplexNonlinearUtilityFunction(UtilityFunction):
         ufuns: Iterable[UtilityFunction],
         combination_function=Callable[[Iterable[UtilityValue]], UtilityValue],
         name=None,
-        reserved_value: Optional[UtilityValue] = 0.0,
+        reserved_value: Optional[UtilityValue] = INVALID_UTILITY,
         ami: AgentMechanismInterface = None,
     ):
         super().__init__(name=name, reserved_value=reserved_value, ami=ami)
@@ -2195,7 +2199,7 @@ class IPUtilityFunction(UtilityFunction):
         distributions: Iterable["UtilityDistribution"] = None,
         issue_names: Iterable[str] = None,
         name=None,
-        reserved_value: Optional[UtilityValue] = 0.0,
+        reserved_value: Optional[UtilityValue] = INVALID_UTILITY,
         ami: AgentMechanismInterface = None,
     ):
         super().__init__(name=name, reserved_value=reserved_value, ami=ami)
@@ -2547,7 +2551,7 @@ def normalize(
     outcomes: Collection[Outcome],
     rng: Tuple[float, float] = (0.0, 1.0),
     epsilon: float = 1e-6,
-    infeasible_cutoff: Optional[float] = -1000.0,
+    infeasible_cutoff: Optional[float] = INVALID_UTILITY,
 ) -> UtilityFunction:
     """Normalizes a utility function to the range [0, 1]
 
@@ -2600,7 +2604,7 @@ def normalize(
                     ami=ufun.ami,
                 )
     scale = (rng[1] - rng[0]) / (mx - mn)
-    r = scale * (ufun.reserved_value - mn) if ufun.reserved_value else 0.0
+    r = scale * (ufun.reserved_value - mn) if ufun.reserved_value is not None else 0.0
     if infeasible_cutoff is not None:
         return ComplexNonlinearUtilityFunction(
             ufuns=[ufun],