+++ HISTORY.rst
@@ -1,6 +1,15 @@
 History
 =======
 
+Release 0.2.22
+--------------
+
+- documentation update
+- changing the reserved_value calculation for miners and GreedyFactoryManager.
+- unifying the INVALID_UTILITY value used by all agents/negotiators to the value defined in utilities.INVALID_UTILITY
+- Aspiration negotiator uses a minimum that is always at least as good as max(reserved_value, INVALID_UTILITY)
+- Added reserved_value parameter to GreedyFactoryManager that allows for control of the reserved value used in all its ufuns.
+
 Release 0.2.21
 --------------